Add a config setting for client exceptions
Exceptions are good for a developer, but can hide errors for end users.

Example, user uploads a file that is too large for clamav (> 25Mb). ClamAV will close the
stream once it hits it's own configured limit. Throwing an exception in this case will show
the user a 500 internal error, instead of showing a useful message (the file size
validation error).

Resolves sunspikes/clamav-validator#50